+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [Unreleased]
 
 ### Added
+- Added a AwsSigningHttpClientFactory to create a signing HTTP client ([#314](https://github.com/opensearch-project/opensearch-php/pull/314))
 ### Changed
 ### Deprecated
 ### Removed